Remove static uri's

+++ STracker/src/src/stracker/asynchttp/AbstractAsyncHttp.java
@@ -6,11 +6,14 @@ import java.net.URL;
 import java.net.URLEncoder;
 import java.util.HashMap;
 import java.util.Map.Entry;
+
+import src.stracker.R;
 import src.stracker.STrackerApp;
+import src.stracker.utils.Utils;
 import HawkClient.*;
+import android.app.Activity;
 import android.app.ProgressDialog;
 import android.content.Context;
-import android.util.Log;
 
 import com.loopj.android.http.AsyncHttpClient;
 import com.loopj.android.http.AsyncHttpResponseHandler;
@@ -52,21 +55,23 @@ public abstract class AbstractAsyncHttp {
 	}
 	
 	public void authorizedGet(String url,STrackerApp app) {
+		if(!Utils.checkLogin((Activity)_context, app)) return;
 		//Waiting message
 		_dialog.setMessage("loading...");
 		_dialog.show();
-		Log.d("header", getAuthorizationHeader("GET",url, app, null));
 		_client.addHeader("Authorization", getAuthorizationHeader("GET",url, app, null));
 		_client.get(url, _handler);
 	}
 	
 	public void authorizedPost(String url, STrackerApp app, HashMap<String, String> params){
+		if(!Utils.checkLogin((Activity)_context, app)) return;
 		PostParams postParams = buildRequestBody(params);
 		_client.addHeader("Authorization", getAuthorizationHeader("POST",url, app, postParams.payload));
 		_client.post(url,postParams.params,_handler);
 	}
 	
 	public void authorizedDelete(String url, STrackerApp app){
+		if(!Utils.checkLogin((Activity)_context, app)) return;
 		_client.addHeader("Authorization", getAuthorizationHeader("DELETE",url, app, null));
 		_client.delete(url,_handler);
 	}
@@ -101,7 +106,7 @@ public abstract class AbstractAsyncHttp {
 		Long time = System.currentTimeMillis();
 		Long timestamp = time / 1000L;
 		String nonce = "NONO";//HawkClient.generateNonce();
-		HawkCredentials credentials = new HawkCredentials(app.getFbUser().getId(), app.getHawkKey());
+		HawkCredentials credentials = new HawkCredentials(app.getFbUser().getId(), _context.getString(R.string.hawk_key));
 		String header = "";
 		try {
 			if(method.equals("POST")){
